Refactored to include routes
