+++ lib/app.js
@@ -1,12 +1,13 @@
 const parsePath = require('./helpers/parsePath');
-const puppies = require('./puppies');
+const notFound = require('./helpers/notFound');
+const puppies = require('./routes/puppies');
 
 const routes = {
   'puppies': puppies
 };
 
 function app(req, res) {
-  const url = parsePath(req.url);
+  const url = parsePath(req.url) || notFound;
   req.query = url.query;
   req.params = url.params;
