+++ lib/helpers/notFound.js
@@ -1,7 +1,7 @@
 function notFound(req, res) {
   res.statusCode = 404;
-  res.statusMessage = `CANNOT GET ${req.url}`;
-  res.end();
+  res.statusMessage = `CANNOT ${req.method} ${req.url}`;
+  res.end(res.statusMessage);
 }
 
 module.exports = notFound;
