+++ lib/routes/puppies.js
@@ -1,4 +1,4 @@
-const connect = require('./helpers/connect');
+const connect = require('../helpers/connect');
 const ObjectId = require('mongodb').ObjectId;
 
 function puppies(req, res) {
